move: infrastructure 디렉토리로 이동
