+++ src/main/java/com/nexters/goalpanzi/infrastructure/auth/RefreshTokenRepositoryImpl.java
@@ -1,0 +1,36 @@
+package com.nexters.goalpanzi.infrastructure.auth;
+
+import com.nexters.goalpanzi.domain.auth.RefreshTokenRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.stereotype.Repository;
+
+import java.time.Duration;
+
+@Repository
+@RequiredArgsConstructor
+public class RefreshTokenRepositoryImpl implements RefreshTokenRepository {
+
+    private static final String REFRESH_TOKEN_POSTFIX = ":refresh_token";
+
+    private final RedisTemplate<String, String> redisTemplate;
+
+    public void save(String altKey, String refreshToken, long ttl) {
+        String key = makeKey(altKey);
+        redisTemplate.opsForValue().set(key, refreshToken, Duration.ofMillis(ttl));
+    }
+
+    public String find(String altKey) {
+        String key = makeKey(altKey);
+        return redisTemplate.opsForValue().get(key);
+    }
+
+    public Boolean delete(String altKey) {
+        String key = makeKey(altKey);
+        return redisTemplate.delete(key);
+    }
+
+    private String makeKey(String altKey) {
+        return altKey + REFRESH_TOKEN_POSTFIX;
+    }
+}
